Refactor Load, Maze, Network actions hpp=>h/cpp
