+++ src/openrct2-ui/input/KeyboardShortcut.cpp
@@ -18,7 +18,7 @@
 #include <openrct2/Game.h>
 #include <openrct2/Input.h>
 #include <openrct2/OpenRCT2.h>
-#include <openrct2/actions/LoadOrQuitAction.hpp>
+#include <openrct2/actions/LoadOrQuitAction.h>
 #include <openrct2/actions/SetCheatAction.hpp>
 #include <openrct2/audio/audio.h>
 #include <openrct2/config/Config.h>
