+++ src/openrct2/actions/GameActionRegistration.cpp
@@ -33,10 +33,10 @@
 #include "LargeSceneryPlaceAction.h"
 #include "LargeSceneryRemoveAction.h"
 #include "LargeScenerySetColourAction.h"
-#include "LoadOrQuitAction.hpp"
-#include "MazePlaceTrackAction.hpp"
-#include "MazeSetTrackAction.hpp"
-#include "NetworkModifyGroupAction.hpp"
+#include "LoadOrQuitAction.h"
+#include "MazePlaceTrackAction.h"
+#include "MazeSetTrackAction.h"
+#include "NetworkModifyGroupAction.h"
 #include "ParkEntranceRemoveAction.hpp"
 #include "ParkMarketingAction.hpp"
 #include "ParkSetDateAction.hpp"
