+++ src/openrct2/network/NetworkBase.cpp
@@ -14,8 +14,8 @@
 #include "../GameStateSnapshots.h"
 #include "../OpenRCT2.h"
 #include "../PlatformEnvironment.h"
-#include "../actions/LoadOrQuitAction.hpp"
-#include "../actions/NetworkModifyGroupAction.hpp"
+#include "../actions/LoadOrQuitAction.h"
+#include "../actions/NetworkModifyGroupAction.h"
 #include "../actions/PeepPickupAction.hpp"
 #include "../core/Guard.hpp"
 #include "../core/Json.hpp"
